Implement edge-to-edge
In order to better accommodate this change, I needed to do a few things first. - The main layout was abstracted into a ToolbarFragment so the MaterialToolbar's lift on scroll functionality would work best (I've found that it frequently has issues with anything that isn't a recycler view and if you swap between multiple different scrolling lists with a single toolbar). The ToolbarFragment handles the initial basic setup of the screen and gives you access to the toolbar and the ability to replace the frame layout inside. This allows it to be reused for every screen in the app. - As a result of the previous point, I converted the language list into a recycler view. - InsetUtil was created to easily add inset listeners with the max insets created on each side of a view based on the system bar and display cutout insets.

The app now displays properly with edge to edge enabled while using the shortEdges window mode which allows the UI to appear around display cutout insets when in certain orientations.
