+++ src/android/RHVoice-core/src/main/AndroidManifest.xml
@@ -12,7 +12,7 @@
     android:label="@string/app_name"
     android:icon="@mipmap/ic_launcher"
     android:roundIcon="@mipmap/ic_launcher_round"
-    android:theme="@style/Theme.Material3.DayNight">
+    android:theme="@style/Theme.RHVoice">
     <service
         android:label="@string/app_name"
         android:name=".RHVoiceService"
