+++ src/android/RHVoice-core/src/main/java/com/github/olga_yakovleva/rhvoice/android/SettingsActivity.java
@@ -16,8 +16,11 @@
 
 package com.github.olga_yakovleva.rhvoice.android;
 
+import android.content.Context;
+import android.content.Intent;
 import android.os.Bundle;
 
+import androidx.activity.EdgeToEdge;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.preference.Preference;
 import androidx.preference.PreferenceFragmentCompat;
@@ -26,6 +29,7 @@ import androidx.preference.PreferenceScreen;
 public final class SettingsActivity extends AppCompatActivity implements PreferenceFragmentCompat.OnPreferenceStartScreenCallback, PreferenceFragmentCompat.OnPreferenceDisplayDialogCallback {
     @Override
     public void onCreate(Bundle state) {
+        EdgeToEdge.enable(this);
         super.onCreate(state);
         setContentView(R.layout.frame);
         if (state == null) {
@@ -39,8 +43,8 @@ public final class SettingsActivity extends AppCompatActivity implements Prefere
         if (key == null)
             throw new IllegalStateException();
         Bundle args = new Bundle();
-        args.putString(SettingsFragment.ARG_LANGUAGE_KEY, key);
-        SettingsFragment frag = new SettingsFragment();
+        args.putString(SettingsListFragment.ARG_LANGUAGE_KEY, key);
+        SettingsListFragment frag = new SettingsListFragment();
         frag.setArguments(args);
         getSupportFragmentManager().beginTransaction().replace(R.id.frame, frag, key).addToBackStack(null).commit();
         return true;
@@ -56,5 +60,10 @@ public final class SettingsActivity extends AppCompatActivity implements Prefere
         f.setTargetFragment(caller, 0);
         f.show(getSupportFragmentManager(), ProsodyPreferenceDialogFragment.TAG);
         return true;
+    }
+
+    public static void show(Context context) {
+        Intent intent = new Intent(context, SettingsActivity.class);
+        context.startActivity(intent);
     }
 }
