+++ src/android/RHVoice-core/src/main/res/layout/voice_list_item.xml
@@ -4,7 +4,7 @@
     android:layout_height="wrap_content"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:orientation="horizontal"
-    android:padding="8dp">
+    android:padding="16dp">
   <LinearLayout
       android:layout_width="0dp"
       android:layout_height="wrap_content"
