+++ src/android/RHVoice-core/src/main/res/values/themes.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <style name="Theme.RHVoice" parent="Theme.Material3.DayNight.NoActionBar" />
+</resources>
